requirements.txt
adal==1.2.4
azure-common==1.1.25
azure-eventhub==1.3.3
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
confluent-kafka==1.5.0
cryptography==3.1.1
idna==2.10
isodate==0.6.0
msrest==0.6.19
msrestazure==0.6.4
mysql-connector==2.2.9
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
uamqp==1.2.10
urllib3==1.25.10
